requirements.txt
django==1.10.0
appdirs==1.4.0
argparse==1.2.1
decorator==3.4.0
gmusicapi==7.0.0
httplib2==0.9
mock==1.0.1
mutagen==1.26
oauth2client==1.4.1
proboscis==1.2.6.0
protobuf==2.6.1
pyasn1==0.1.7
pyasn1-modules==0.0.5
python-dateutil==2.2
python-mpd2==0.5.3
requests==2.10.0
rsa==3.1.4
six==1.8.0
validictory==1.0.0a2
wsgiref==0.1.2
django-debug-toolbar==1.4
django-bootstrap3==7.0.1
RPi.GPIO==0.6.2
djangorestframework==3.5.0
django-sortedm2m==1.3.2
model-mommy==1.3.0
